Move file handlers from api to utils
